refactor: split CSSOM parser module
